requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
redis>=4.5.2,<4.5.5  
celery[redis]==5.3.1
python-jose[cryptography]==3.3.0
langchain==0.0.348
transformers==4.35.2
torch==2.5.1
pytest==7.4.0
pytest-django==4.5.2
psutil==5.9.5
docker==6.1.3
scikit-learn==1.3.0
numpy==1.24.3
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.1


